Code cleanup

+++ lib/djvCore/StringInline.h
@@ -333,10 +333,10 @@ namespace djv
                     ;
 
                 // Add up the digits.
-                int tens = 1;
+                int64_t tens = 1;
                 for (int i = int(end) - 1; i >= 0; --i, tens *= 10)
                 {
-                    out += static_cast<int64_t>((s[i] - 48) * tens);
+                    out += (static_cast<int64_t>(s[i]) - 48) * tens;
                 }
 
                 // Apply the sign.
@@ -351,10 +351,10 @@ namespace djv
                 out = 0;
 
                 // Add up the digits.
-                int tens = 1;
+                size_t tens = 1;
                 for (int i = int(size) - 1; i >= 0; --i, tens *= 10)
                 {
-                    out += static_cast<size_t>((s[i] - 48) * tens);
+                    out += (static_cast<size_t>(s[i]) - 48) * tens;
                 }
             }
